requirements.txt
streamlit==1.32.0
sentence-transformers==5.1.1
transformers==4.57.0
faiss-cpu==1.12.0
torch==2.8.0
numpy==2.1.3
accelerate==1.11.0
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2